requirements.txt
annotated-types==0.6.0
anyio==4.2.0
attrs==23.2.0
certifi==2023.11.17
click==8.1.7
distro==1.9.0
fastapi==0.108.0
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
idna==3.6
inflection==0.5.1
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
openai==1.7.0
openapi==1.1.0
pydantic==2.5.3
pydantic_core==2.14.6
python-dotenv==1.0.0
PyYAML==6.0.1
referencing==0.32.1
rpds-py==0.16.2
sniffio==1.3.0
starlette==0.32.0.post1
tqdm==4.66.1
typing_extensions==4.9.0
uvicorn==0.25.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
